Added support for Accelerometer Sensor
